+++ demisto_sdk/commands/content_graph/strict_objects/common.py
@@ -1,5 +1,5 @@
 from abc import ABC
-from typing import Any, Dict, Optional, Sequence
+from typing import Any, Dict, Optional, Sequence, List, Union
 
 import pydantic
 from pydantic import BaseModel, Extra, validator
